+++ src/mapping/stability_measures_accumulator.jl
@@ -12,10 +12,11 @@ Distributions.pdf(::EverywhereUniform, u) = one(eltype(u))
 
 A special data structure that allows mapping initial conditions to attractors
 while _at the same time_ calculating many stability measures in the most efficient
-way possible. `mapper` is any instance of an [`AttractorMapper`](@ref)
-that implements the `id = mapper(u0)` syntax. This functionality was developed
-as part of [Morr2026](@cite).
+way possible. `mapper` is any instance of an [`AttractorMapper`](@ref),
+although for [`AttractorsViaFeaturizing`](@ref) the convergence times won't make sense.
 
+This functionality was developed as part of [Morr2026](@cite) and has now been extended
+to work for any `AttractorMapper` current or future.
 The accummulator records several measures of stability (or resilience) defined
 in [Morr2026](@cite), and a few more related and derived shortly after, see list below.
 However, it also allows computing any additional user-defined quantifier that is
@@ -225,7 +226,7 @@ function convergence_time(accumulator::StabilityMeasuresAccumulator)
 end
 
 # Function to update the accumulator with a new state
-function (accumulator::StabilityMeasuresAccumulator)(u0; show_progress = false)
+function (accumulator::StabilityMeasuresAccumulator)(u0)
     id = accumulator.mapper(u0)
     push!(accumulator.u0s, u0)
     push!(accumulator.bs, id)
@@ -243,15 +244,16 @@ See [`StabilityMeasuresAccumulator`](@ref) for more.
 function finalize_accumulator(accumulator::StabilityMeasuresAccumulator; discrete_time_Δt = 1.0, probability_cutoff = 0.0)
     ds = referenced_dynamical_system(accumulator)
     attractors = extract_attractors(accumulator.mapper)
-    ids = vcat(collect(keys(attractors)), -1)
-    js = 1:length(ids)
-    ids_to_js = Dict(id => j for (j, id) in enumerate(ids))
     u0s = accumulator.u0s
     bs = accumulator.bs
     cts = accumulator.cts
+    ids = vcat(collect(keys(attractors)), -1) # Output stability also for attractors with no basin points
+    js = 1:length(ids)
+    ids_to_js = Dict(id => j for (j, id) in enumerate(ids))
     N = length(u0s)
     N == 0 && error("No initial conditions have been processed. Cannot finalize accumulator.")
 
+    # make distance matrix attractors <-> u0s
     d = zeros(length(u0s), length(js))
     for i in 1:length(u0s)
         for j in js
@@ -263,12 +265,12 @@ function finalize_accumulator(accumulator::StabilityMeasuresAccumulator; discret
         end
     end
 
-    if (isa(accumulator.mapper, AttractorsViaProximity) && accumulator.mapper.ε != nothing)
+    # convergence paces
+    if (isa(accumulator.mapper, AttractorsViaProximity) && !isnothing(accumulator.mapper.ε))
         ε = accumulator.mapper.ε
     else
         ε = 0.0
     end
-
     cps = zeros(length(cts))
     for i in 1:length(cts)
         j = ids_to_js[bs[i]]
@@ -284,6 +286,7 @@ function finalize_accumulator(accumulator::StabilityMeasuresAccumulator; discret
       cps .*= discrete_time_Δt
     end
 
+    # now perform the bulk of nonlocal stability estimation
     basin_frac = Dict(id => 0.0 for id in ids)
     basin_stab = Dict(id => 0.0 for id in ids)
     finite_time_basin_stab = Dict(id => 0.0 for id in ids)
@@ -312,18 +315,15 @@ function finalize_accumulator(accumulator::StabilityMeasuresAccumulator; discret
         end
 
         mean_convergence_time[id] += w * ct / N
-
         mean_convergence_pace[id] += w * cp / N
 
         if w > 0.0
             maximal_convergence_time[id] = max(
                 maximal_convergence_time[id], ct
             )
-
             maximal_noncritical_shock_magnitude[id] = max(
                 maximal_noncritical_shock_magnitude[id], d[i, j]
             )
-
             maximal_convergence_pace[id] = max(
                 maximal_convergence_pace[id], cp
             )
@@ -361,8 +361,9 @@ function finalize_accumulator(accumulator::StabilityMeasuresAccumulator; discret
             )
             for id in ids
     )
-    minimal_critical_shock_magnitude[-1] = NaN # no critical shock for -1 attractor
-    
+    if -1 ∈ ids
+        minimal_critical_shock_magnitude[-1] = NaN # no critical shock for -1 attractor
+    end
     # linear measures
     characteristic_return_time = Dict(id => NaN for id in ids)
     reactivity = Dict(id => NaN for id in ids)
@@ -466,4 +467,22 @@ function weighted_median(
         end
     end
     return float(vals[p[end]])
+end
+
+# Extension function to allow accumulator to work also with non-single-u0 mappers
+function basins_fractions_grouped(accumulator::StabilityMeasuresAccumulator, ics, progress::Progress, labels)
+    # note that ics is always collected into vector, cascaded by `basins_fraction`.
+    # Here we'll call the same function with the stored mapper, cheating the system
+    fs, _labels = basins_fractions(accumulator.mapper, ics; show_progress = progress.core.enabled)
+    if !isempty(labels)
+        labels .= @view(_labels[1:length(labels)])
+    end
+    # and then we'll use the found labels and store the `ics`. The downside is that
+    # we can't have a proper convergence time, but that's okay
+    for i in eachindex(ics)
+        push!(accumulator.u0s, ics[i])
+        push!(accumulator.bs, _labels[i])
+        push!(accumulator.cts, NaN)
+    end
+    return fs
 end